name moveDraggableAction
